moved files in new dir, edited the paths
