+++ signaturenDruck/js/K10plus/dataExtractK10plus.js
