+++ signaturenDruck/js/K10plus/loadDataFromFileK10plus.js
@@ -4,9 +4,9 @@ const config = remote.getGlobal('config')
 // requires the dataExtractK10plus-module
 const DataExtractK10plus = require('./dataExtractK10plus.js')
 // requires the shelfmark class
-const Shelfmark = require('./shelfmark.js')
-const getLabelSize = require('./getLabelSize.js')
-const createMultipleLines = require('./createMultipleLines.js')
+const Shelfmark = require('../shelfmark.js')
+const getLabelSize = require('../getLabelSize.js')
+const createMultipleLines = require('../createMultipleLines.js')
 
 module.exports = function (allLines) {
   let obj = {
