+++ signaturenDruck/js/preK10plus/dataExtractOld.js
